fix: the game will refuse to load if the save version is newer
